requirements.txt
# Core RAG Dependencies
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.24.0

# FastAPI Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
python-magic>=0.4.27
chardet>=5.2.0

# Web Interface
streamlit>=1.29.0
streamlit-chat>=0.1.1

# Configuration Management
hydra-core>=1.3.2
pydantic>=2.5.0
pyyaml>=6.0.1

# Data Processing
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0

# HTTP Client for Ollama
requests>=2.31.0
httpx>=0.25.0

# Logging and Monitoring
loguru>=0.7.2
rich>=13.7.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=23.12.0
isort>=5.13.0
flake8>=6.1.0
mypy>=1.8.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
click>=8.1.7
timm==1.0.19